+++ angel_serialize_generator/test/models/author.d.ts
@@ -1,29 +1,28 @@
-/// <reference path="../book.d.ts" />
 // GENERATED CODE - DO NOT MODIFY BY HAND
 declare module 'angel_serialize_generator' {
   interface Author {
     id?: string;
     name: string;
     age: number;
-    books?: Book[];
-    newest_book?: Book;
+    books?: any[];
+    newest_book?: any;
     created_at?: any;
     updated_at?: any;
   }
   interface Library {
     id?: string;
-    collection?: LibraryCollection;
+    collection: LibraryCollection;
     created_at?: any;
     updated_at?: any;
   }
   interface LibraryCollection {
-    [key: string]: Book;
+    [key: string]: any;
   }
   interface Bookmark {
     id?: string;
-    history?: number[];
+    history: number[];
     page: number;
-    comment?: string;
+    comment: string;
     created_at?: any;
     updated_at?: any;
   }