+++ angel_serialize_generator/test/models/book.d.ts
@@ -2,11 +2,11 @@
 declare module 'angel_serialize_generator' {
   interface Book {
     id?: string;
-    author?: string;
-    title?: string;
-    description?: string;
-    page_count?: number;
-    not_models?: number[];
+    author: string;
+    title: string;
+    description: string;
+    page_count: number;
+    not_models: number[];
     camelCase?: string;
     created_at?: any;
     updated_at?: any;