+++ angel_serialize_generator/test/models/with_enum.dart
@@ -6,10 +6,13 @@ part 'with_enum.g.dart';
 
 @serializable
 abstract class _WithEnum {
+  @DefaultsTo(WithEnumType.b)
   WithEnumType get type;
 
+  @nullable
   List<int> get finalList;
 
+  @nullable
   Uint8List get imageBytes;
 }